refactor: mobile-wrapper resize hook 분리
